Run example in simulator without xcode
Adds a script `example/bin/run-ios` that runs the example app in the iOS
simulator, without having to start XCode first. Also adds a helper
script, `example-bin/tail-ios-sim`, that tails the simulator log (what
you would see in the XCode debuge console).

This relies on code in react-native that didn't make it into the 0.18
release. For this reason, set react-native dependency to a specific commit in
package.json.