Fix android logging
For me, this change is necessary to see any output from the android
logger.

In my experiments I found out that using conch for any process that
doesn't terminate would never print anything to the screen, So `(ls)`
would work fine but `(ping "www.google.com")` would never flush to the
screen.

I made a note to research this more and see if I can determine the
problem before I submit a bug on the conch repo. It could possibly be
platform-specific, as I am on OS X.

Hopefully I can get this figured out and fixed so we could go back to
conch once the `process-line` functionality is done