Run RN packager in a tmpdir
Rather than running the React Native packager in the appdir itself,
create a Boot tmpdir, copy the required node_modules into it, and run
the packager there.

By running from a tmpdir, we can construct exactly the directory
structure that the packager expects from a standard JavaScript
project without needing to structure the boot fileset that way. This
means we can play nicely with other boot tasks that the project might
define -- such as other CLJS builds in the same project (ie
non-React-Native code that shares a codebase with our React Native
code).