Use transformer absolute path for bundle task
I don't know why, but on a few occasions I've seen the bundle task fail,
complaining that it can't find the transformer, if given a relative path
instead of an absolute one. This doesn't appear to happen to the
start-rn-packager task (well, at least, I haven't seen it happen there
so far).