Remove reference to deprecated commands
The commands run-in-simulator, print-ios-log and print-android-log
should not be used anymore. Instead, run `react-native run-ios`,
`react-native log-ios` and its android equivalents directly from
the command line, in a separate terminal window.

The reason for this change is that the tools bundled with react-native
are more reliable and more consistent with the default react-native
experience.